feat: add hashHistory and change hashRouter to Router
